Extract date/time pickers into components
Add reusable DatePickerField and TimePickerField composables and replace the inline implementations in screens. Two new files (DatePickerField.kt, TimePickerField.kt) provide a DatePickerDialog-based date field (YYYY-MM-DD) and a 24-hour TimePicker dialog (HH:mm). Removed duplicate picker code from CreateJobScreen and updated imports in CreateJobScreen and EditJobScreen to use the new components, simplifying UI code and enabling reuse.
