+++ app/src/main/java/com/moviles/jobmatch/ui/screens/company/CreateJobScreen.kt
@@ -22,12 +22,11 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.moviles.jobmatch.ui.components.DatePickerField
 import com.moviles.jobmatch.ui.components.SectionHeader
-import java.util.Calendar
-import java.util.TimeZone
+import com.moviles.jobmatch.ui.components.TimePickerField
 
 val JobMatchBlue = Color(0xFF2196F3)
 
@@ -439,126 +438,6 @@ fun CreateJobScreen(
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun DatePickerField(
-    label: String,
-    value: String,
-    onDateSelected: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    var showDialog by remember { mutableStateOf(false) }
-    val state = rememberDatePickerState()
-
-    Box(modifier = modifier) {
-        OutlinedTextField(
-            value = value,
-            onValueChange = {},
-            label = { Text(label) },
-            placeholder = { Text("Selecciona una fecha", color = Color.Gray) },
-            trailingIcon = {
-                Icon(Icons.Outlined.CalendarToday, contentDescription = null, tint = Color.Gray)
-            },
-            readOnly = true,
-            modifier = Modifier.fillMaxWidth(),
-            shape = RoundedCornerShape(8.dp)
-        )
-        // Overlay that makes the whole field tappable while keeping readOnly
-        Box(modifier = Modifier.matchParentSize().clickable { showDialog = true })
-    }
-
-    if (showDialog) {
-        DatePickerDialog(
-            onDismissRequest = { showDialog = false },
-            confirmButton = {
-                TextButton(onClick = {
-                    state.selectedDateMillis?.let { millis ->
-                        val cal = Calendar.getInstance(TimeZone.getTimeZone("UTC"))
-                        cal.timeInMillis = millis
-                        onDateSelected(
-                            "%04d-%02d-%02d".format(
-                                cal.get(Calendar.YEAR),
-                                cal.get(Calendar.MONTH) + 1,
-                                cal.get(Calendar.DAY_OF_MONTH)
-                            )
-                        )
-                    }
-                    showDialog = false
-                }) { Text("Aceptar") }
-            },
-            dismissButton = {
-                TextButton(onClick = { showDialog = false }) { Text("Cancelar") }
-            }
-        ) {
-            DatePicker(state = state)
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun TimePickerField(
-    label: String,
-    value: String,
-    onTimeSelected: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    var showDialog by remember { mutableStateOf(false) }
-    val state = rememberTimePickerState(is24Hour = true)
-
-    Box(modifier = modifier) {
-        OutlinedTextField(
-            value = value,
-            onValueChange = {},
-            label = { Text(label) },
-            placeholder = { Text("HH:mm", color = Color.Gray) },
-            trailingIcon = {
-                Icon(Icons.Outlined.Schedule, contentDescription = null, tint = Color.Gray)
-            },
-            readOnly = true,
-            modifier = Modifier.fillMaxWidth(),
-            shape = RoundedCornerShape(8.dp)
-        )
-        Box(modifier = Modifier.matchParentSize().clickable { showDialog = true })
-    }
-
-    if (showDialog) {
-        Dialog(onDismissRequest = { showDialog = false }) {
-            Card(
-                shape = RoundedCornerShape(16.dp),
-                colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface)
-            ) {
-                Column(
-                    modifier = Modifier.padding(24.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    Text(
-                        text = label,
-                        style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.SemiBold,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(bottom = 20.dp)
-                    )
-                    TimePicker(state = state)
-                    Row(
-                        modifier = Modifier.fillMaxWidth().padding(top = 8.dp),
-                        horizontalArrangement = Arrangement.End
-                    ) {
-                        TextButton(onClick = { showDialog = false }) { Text("Cancelar") }
-                        Spacer(modifier = Modifier.width(8.dp))
-                        TextButton(onClick = {
-                            onTimeSelected(
-                                "%02d:%02d".format(state.hour, state.minute)
-                            )
-                            showDialog = false
-                        }) { Text("Aceptar") }
-                    }
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun ChipItem(text: String, onRemove: () -> Unit) {
